requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
pytest-cov==4.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
anthropic==0.18.1
cohere==4.48 